ai_accident_detection/requirements.txt
# Core dependencies
ultralytics==8.0.196          # YOLOv8
opencv-python==4.8.1.78       # Video processing
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2

# API & Networking
requests==2.32.2
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
Pillow==10.0.1

# Logging
loguru==0.7.2